SchackSpel/requirements.txt
python-chess
playsound
